Fix an issue in EnterAttributeAddChosenSubtypeEffect that caused clones of Metallic Mimic and Adaptive Automaton to have the chosen type.
Also add a continuous type setting effect to Metallic Mimic so it continues to work as expected after the fix.
