+++ Mage.Sets/src/mage/cards/a/AdaptiveAutomaton.java
@@ -3,10 +3,9 @@ package mage.cards.a;
 
 import java.util.UUID;
 import mage.MageInt;
-import mage.abilities.Ability;
+import mage.abilities.effects.common.continuous.AddChosenSubtypeEffect;
 import mage.abilities.common.AsEntersBattlefieldAbility;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.ContinuousEffectImpl;
 import mage.abilities.effects.common.ChooseCreatureTypeEffect;
 import mage.abilities.effects.common.continuous.BoostAllOfChosenSubtypeEffect;
 import mage.abilities.effects.common.enterAttribute.EnterAttributeAddChosenSubtypeEffect;
@@ -15,8 +14,6 @@ import mage.cards.CardSetInfo;
 import mage.constants.*;
 import mage.filter.common.FilterCreaturePermanent;
 import mage.filter.predicate.permanent.ControllerPredicate;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
 
 /**
  * @author nantuko
@@ -42,7 +39,7 @@ public final class AdaptiveAutomaton extends CardImpl {
         ability.addEffect(new EnterAttributeAddChosenSubtypeEffect());
         this.addAbility(ability);
 
-        this.addAbility(new SimpleStaticAbility(Zone.BATTLEFIELD, new AdaptiveAutomatonAddSubtypeEffect()));
+        this.addAbility(new SimpleStaticAbility(Zone.BATTLEFIELD, new AddChosenSubtypeEffect()));
         // Other creatures you control of the chosen type get +1/+1.
         this.addAbility(new SimpleStaticAbility(Zone.BATTLEFIELD, new BoostAllOfChosenSubtypeEffect(1, 1, Duration.WhileOnBattlefield, filter, true)));
     }
@@ -57,31 +54,3 @@ public final class AdaptiveAutomaton extends CardImpl {
     }
 }
 
-class AdaptiveAutomatonAddSubtypeEffect extends ContinuousEffectImpl {
-
-    public AdaptiveAutomatonAddSubtypeEffect() {
-        super(Duration.WhileOnBattlefield, Layer.TypeChangingEffects_4, SubLayer.NA, Outcome.Benefit);
-        staticText = "{this} is the chosen type in addition to its other types";
-    }
-
-    public AdaptiveAutomatonAddSubtypeEffect(final AdaptiveAutomatonAddSubtypeEffect effect) {
-        super(effect);
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        Permanent permanent = game.getPermanent(source.getSourceId());
-        if (permanent != null) {
-            SubType subType = ChooseCreatureTypeEffect.getChoosenCreatureType(permanent.getId(), game);
-            if (subType != null && !permanent.hasSubtype(subType, game)) {
-                permanent.getSubtype(game).add(subType);
-            }
-        }
-        return true;
-    }
-
-    @Override
-    public AdaptiveAutomatonAddSubtypeEffect copy() {
-        return new AdaptiveAutomatonAddSubtypeEffect(this);
-    }
-}
